requirements.txt
PyYAML==3.10
beautifulsoup4==4.3.1
boto==2.12.0
pyimgur==0.5.2
requests==1.2.3
wsgiref==0.1.2
